Interim commit to progress monitoring code - not yet working

+++ translator/sri/testing/processor.py
@@ -4,17 +4,20 @@ Utility module to support SRI Testing harness background processing
 The module launches the SRT Testing test harness using the Python 'multiprocessor' library.
 See https://docs.python.org/3/library/multiprocessing.html?highlight=multiprocessing for details.
 """
-from multiprocessing import Process
-from multiprocessing.context import BaseContext
-from sys import platform, stdout, stderr
+from sys import platform, stderr
 
-from typing import Optional, Union
+from multiprocessing import Process, Pipe
+from multiprocessing.context import BaseContext
+from multiprocessing.connection import Connection
+
+from typing import Optional, Union, Callable, Generator
 import multiprocessing as mp
 from queue import Empty
 from subprocess import (
-    run,
-    # PIPE,
-    # STDOUT,
+    # run,
+    Popen,
+    PIPE,
+    STDOUT,
     CompletedProcess,
     CalledProcessError,
     TimeoutExpired
@@ -39,10 +42,16 @@ else:
     PWD_CMD = "pwd"
 
 
-def _worker_process(lock: mp.Lock, queue: mp.Queue, command_line: str):
+def _worker_process(
+        pipe: mp.Pipe,
+        lock: mp.Lock,
+        queue: mp.Queue,
+        command_line: str
+):
     """
     Wrapper for a background worker process which runs a specified command.
 
+    :param pipe: Pipe, process pipe between the Worker and caller
     :param lock: Lock, process lock used to synchronize background output
     :param queue: Queue, used for communication of worker process to caller
     :param command_line: str, the worker process command to execute
@@ -63,71 +72,40 @@ def _worker_process(lock: mp.Lock, queue: mp.Queue, command_line: str):
     # the caller, as a point of reference
     queue.put(process_id)
 
-    result: Union[CompletedProcess, CalledProcessError, TimeoutExpired]
+    # result: Union[CompletedProcess, CalledProcessError, TimeoutExpired]
+    return_code: int
+    return_status: str
     try:
         # do the heavy lifting here
-        #     try:
-        #         with Popen(
-        #                 args=cmd,
-        #                 # env=env,
-        #                 bufsize=1,
-        #                 universal_newlines=True,
-        #                 stdout=PIPE,
-        #                 stderr=STDOUT
-        #         ) as proc:
-        #             logger.info(f"run_script({script}) log:")
-        #             for line in proc.stdout:
-        #                 line = line.strip()
-        #                 if line:
-        #                     # propagate the line to the
-        #                     # parent process, via the Queue?
-        #                     queue.put(line)
-        #
-        #     except RuntimeError:
-        #         logger.error(f"run_script({script}) exception: {exc_info()}")
-        #         return -1
-        result = run(
-            command_line,
-            shell=True,
-            check=True,
-            capture_output=True,
-            # stdout=PIPE,
-            # stderr=STDOUT,
-            universal_newlines=True,
-            text=True,
+        with Popen(
+                args=command_line,
+                # env=env,
+                bufsize=1,
+                universal_newlines=True,
+                stdout=PIPE,
+                stderr=STDOUT
+        ) as proc:
+            logger.debug(f"_worker_process(command_line={command_line}) log:\n")
+            for line in proc.stdout:
+                line = line.strip()
+                if line:
+                    # Simple-minded approach of shoving all the
+                    # child output lines into an interprocess pipe
+                    pipe.send(line)
 
-            # running test harnesses may take a very long time
-            # so maybe a bit challenge to set a sensible timeout here
-            # timeout=100,
-        )
+        return_code = proc.returncode
+        return_status = "Worker Process Completed?"
 
-        # TODO: somehow read the stdout stream to parse the PyTest %completion as a progress monitor
-
-    except CalledProcessError as cpe:
-        result = cpe
-
-    # except TimeoutExpired as toe:
-    #     print(str(toe), flush=True, file=stderr)
-    #     result = toe
-
-    lock.acquire()
-    try:
-        print(f"Result: {str(result)}", flush=True, file=stderr)
-
-        return_code = result.returncode
-        print(f"Return Code: {return_code}", flush=True, file=stderr)
-
-        output = result.stdout
-        print(f"Output: {output}", flush=True, file=stdout)
-
-        errors = result.stderr
-        print(f"Errors: {errors}", flush=True, file=stderr)
-
-    finally:
-        lock.release()
+    except RuntimeError as rte:
+        logger.error(f"run_script({command_line}) exception: {str(rte)}")
+        return_code = 1
+        return_status = f"Worker Process Exception: {str(rte)}?"
 
     # propagate the result - successful or not - back to the caller
-    queue.put(result)
+    queue.put(f"Worker Process Return Code: {return_code}\nStatus {return_status}")
+
+    # Job done, close the interprocess pipe?
+    pipe.close()
 
 
 class WorkerProcessException(Exception):
@@ -143,8 +121,7 @@ class WorkerProcess:
         :param timeout: int, worker process data query access timeout (Default: None => queue data access is blocking?)
         """
         self._timeout: Optional[int] = timeout
-        self._result: Optional[Union[CompletedProcess, CalledProcessError, TimeoutExpired, Empty]] = None
-        self._output: Optional[str] = None
+        self._parent_conn: Optional[Connection] = None
         self._ctx: BaseContext = mp.get_context('spawn')
         self._queue = self._ctx.Queue()
         self._lock = self._ctx.Lock()
@@ -164,12 +141,24 @@ class WorkerProcess:
         logger.debug(f"run_command() command: {command_line}")
 
         try:
-            # Sanity check: WorkProcess is a singleton?
+            # Sanity check: Work Process is a singleton?
             # TODO: may need to manage several worker processes, therefore, may need to use multiprocessing Pools? see
             #   https://docs.python.org/3/library/multiprocessing.html?highlight=multiprocessing#module-multiprocessing
             assert not self._process
-            
-            self._process = self._ctx.Process(target=_worker_process, args=(self._lock, self._queue, command_line))
+
+            # We'll access the internal _worker_proces
+            # standard output/error stream, via a Pipe
+            self._parent_conn, child_conn = Pipe()
+
+            self._process = self._ctx.Process(
+                target=_worker_process,
+                args=(
+                    child_conn,
+                    self._lock,
+                    self._queue,
+                    command_line
+                )
+            )
 
             self._process.start()
 
@@ -192,54 +181,20 @@ class WorkerProcess:
 
             logger.warning(f"run_command() command: '{command_line}' raised an exception: {str(ex)}?")
 
-            if self._process:
+            if self._process and self._process.is_alive():
                 self._process.kill()
-                self._process = None
-                self._process_id = 0
-            
-            self._output = f"Background process start-up exception: {str(ex)}?"
 
-    def get_output(self) -> Optional[str]:
-        """
-        Retrieves the raw STDOUT output or (alternately) error messages
-        from the WorkerProcess, when it is complete.
+            self._process = None
+            self._process_id = 0
 
-        :return: Optional[str], output which may be a raw worker process result,
-                                an error message or None if not yet available.
-        """
-        # this method is idempotent: once a non-empty output is
-        # retrieved the first time, it is deemed ached for future access
-        if not self._output:
-            try:
-                self._result = self._queue.get(block=True, timeout=self._timeout)
-                logger.debug(f"WorkerProcess.output() result:\n\t{self._result}")
-            except Empty as empty:
-                logger.debug("Worker Process output is not (yet) available?")
-                self._result = empty
+    def get_output(self) -> Generator:
+        if self._parent_conn:
+            while True:
+                try:
+                    if self._parent_conn.poll(1):
+                        yield self._parent_conn.recv()
+                except EOFError:
+                    return None
 
-            if self._result:
-                if isinstance(self._result, CompletedProcess):
-                    self._output = self._result.stdout  # sending back full raw process standard output
-                    if self._result.returncode != 0:
-                        # A special warning is added to the output?
-                        self._output = "WARNING: Worker Process returned a non-zero return code: " + \
-                                      f"'{str(self._result.returncode)}': \n\t{self._output}"
-                    self._process.join()
-
-                elif isinstance(self._result, Empty) or isinstance(self._result, TimeoutExpired):
-                    # Worker process still running, self.process likely still running and the 'process_id'
-                    # is likely (still) valid. Data 'output' is left None for now? Try again later...
-                    pass
-
-                elif isinstance(self._result, CalledProcessError):
-                    # process aborted by internal error?
-                    self._output = self._result.stdout
-                    if self._process:
-                        self._process.kill()
-                        self._process = None
-                else:
-                    raise WorkerProcessException(
-                        f"ERROR: Unexpected result type encountered from worker process: {type(self._result)}"
-                    )
-            
-        return self._output
+    def close(self):
+        self._process.join()
